Description of patch creation process
I suggest these improvements to the description of how the patch is applied to a map, and also how the patch is created. I hope it would be of value to anyone who wants to understand how the maps in AI Arena are modified or who wants to troubleshoot problems with patched maps.

This is my understanding of the process as received from the discussion in the discord channel. Feel free to use any or all text, as well as correct anything I got wrong.